x1plusd: Dynamically load/register actions from directory
