+++ images/cfw/opt/x1plus/lib/python/x1plus/services/x1plusd/action_types/delay.py
@@ -1,0 +1,18 @@
+"""
+[action-type]
+name=delay
+[end]
+"""
+from ..actions import register_action
+
+import asyncio
+import logging
+
+logger = logging.getLogger(__name__)
+
+@register_action("delay")
+async def _action_delay(handler, subconfig):
+    logger.debug(f"delay action: {subconfig}")
+    if type(subconfig) != int and type(subconfig) != float:
+        raise TypeError(f"delay parameter {subconfig} was not numberish")
+    await asyncio.sleep(subconfig)
