+++ images/cfw/opt/x1plus/lib/python/x1plus/services/x1plusd/action_types/syslog.py
@@ -1,0 +1,14 @@
+"""
+[action-type]
+name=syslog
+[end]
+"""
+from ..actions import register_action
+
+import logging
+
+logger = logging.getLogger(__name__)
+
+@register_action("syslog")
+async def _action_syslog(handler, subconfig):
+    logger.info(f"syslog action: {subconfig}")
